+++ docs/changelog.md
@@ -1,6 +1,8 @@
 # sedrila CHANGELOG
 
-## Version 3.x (upcoming)
+## Version 3.1.0 (upcoming)
+- `student`, `instructor`: Webapp shows task list and workreport in chapter/taskgroup order
+  rather than alphabetical order
 - ...